chore(ds): move component definition - update utils
